relocated styles files because in the pages folder, you need to export a react component, and that is not the case in there. Living and Learning.
